requirements.txt
motor
pydantic
requests
python-dotenv
matplotlib
